properly namespace rash inside hashie
