+++ lib/rash/version.rb
@@ -1,3 +1,7 @@
-class Rash
-  VERSION = '0.2.0'
+require 'hashie/rash'
+
+module Hashie
+  class Rash < Mash
+    VERSION = '0.2.0'
+  end
 end
